Move LabelOptions and SegmentationTask to separate package config
